requirements.txt
gTTS>=2.5.1
langchain>=0.2.6
langchain-community>=0.2.6
langchain-core>=0.2.10
langchain-openai>=0.1.10
matplotlib==3.9.2
openai>=1.35.4
pandas>=2.2.2
pillow>=10.2.0
python-dotenv>=1.0.1
scikit-learn==1.5.1
streamlit>=1.36.0
torch>=2.2.2
torchvision>=0.17.2
transformers>=4.42.4
